more tweaks for the occasional usb halt during ehci_hcd quiesce.  I had a mismatch in the mutex locks from the last pass.  This was resulting in a panic during shutdown independent of whether anything was plugged into USB.  (This was particularly causing firmware updates to fail) and required a download mode step to fix.
This fixes the typo and comments out the kernel 2.6 codepath that I
REplaced.  in the process, found that ehci-sched.c was calling the
removed bit.  Converted that to just handshake instead of handshake and
halt.

upstream kernel completely refactored this codebase with some sort of
polling timing thing but I don't want to peel the onion if i can avoid
it.